En dev - TODOs

+++ src/src-fw/initsession.ts
@@ -9,25 +9,34 @@ type age = {
   lsd: number // jour (EPOCH) de dernière synchro aboutie
 }
 
-export const initSync = async (dbReset: boolean) => {
+export const sessionPhase0 = async (
+  mode: modes, dbReset: boolean, dbName: string
+  // TODO - arguments pour subscriptions et auths en INCOGNITO
+  ) => {
+  
   // Acquisition des souscriptions et documents
   const session = stores.session
   const dataSt = stores.data
   const config = stores.config
 
-  session.setDbName('dbloc')
-  session.setMode(modes.SYNC)
-  if (dbReset)
+  session.setMode(mode)
+  if (session.hasIDB)
+    session.setDbName(dbName)
+  
+  if (session.mode === modes.SYNC && dbReset)
     await IDB.delete(session.dbName)
 
-  const idb = await IDB.open()
+  let idb = session.hasIDB ? await IDB.open() : null
+  let integral = true
+  const jourj = Math.floor(Date.now() / 86400000)
 
-  const age = await idb.getState('age') as age
-  const j = Math.floor(Date.now() / 86400000)
-  const integral = !age.lsd || (j - age.lsd) > config.K.SYNCINCRNBD
+  if (session.mode === modes.SYNC) {
+    const age = await idb.getState('age') as age
+    integral = !age.lsd || (jourj - age.lsd) > config.K.SYNCINCRNBD
+  }
 
   // Inscription en data-store des defs de toutes les sousciptions
-  {
+  if (session.mode !== modes.PLANE) {
     const m = await idb.getSubs()
     for(const [org, mo] of m) {
       for(const [clazz, subs] of mo)
@@ -35,32 +44,35 @@ export const initSync = async (dbReset: boolean) => {
     }
   }
 
-  /* Enregistrement / abonnement au serveur des souscriptions de "session active"
-  Redéfinition possible de title, donnée de url
+  /* Enregistrement en store (et abonnement au serveur) 
+  des souscriptions de "session active".
   Des notifications peuvent désormais parvenir mais sont accumulées et NON traitées:
   la queue de traitement n'est ouverte qu'à la fin de la phase 0.
   */
-  {
+  if (session.mode !== modes.PLANE) {
     const allSubs: Map<string, Subscription> = await idb.getSubscriptions()
     for(const [org, subscription] of allSubs) {
       dataSt.setSubscription(org, subscription)
-      await subscription.subscribe(org, true)
+      await subscription.subscribe(org, true) // Redéfinition possible de title / url
     }
   }
 
-  if (integral) {
+  if (session.mode === modes.SYNC && integral)
     await idb.deleteAllDocs()
-  } else {
+  if (session.hasIDB && !integral)
     await idb.loadAllDocs()
-  }
   
   // synchro immediate (sequentielle) de toutes les souscriptions
-  await dataSt.syncAll()
+  if (session.hasNet)
+    await dataSt.syncAll()
   // Marquage de l'age de l'état de synchro de IDB
-  await idb.putState('age', { lsd: j })
+  if (session.mode === modes.SYNC)
+    await idb.putState('age', { lsd: jourj })
 
   // Fin de phase 1
   session.setPhase(1)
+
   // Les notifications reçues demandant des sync ne sont plus bloquées en queue
-  dataSt.startSyncQueue() 
+  if (session.hasNet)
+    dataSt.startSyncQueue() 
 }
